+++ .travis/install.sh
@@ -42,21 +42,21 @@ if [[ "$(uname -s)" == 'Darwin' ]]; then
             pyenv global 3.5.0
             ;;
         pypy)
-            pyenv install pypy-2.6.0
-            pyenv global pypy-2.6.0
+            pyenv install "pypy-${PYPY_VERSION}"
+            pyenv global "pypy-${PYPY_VERSION}"
             ;;
     esac
     pyenv rehash
     python -m pip install -U --user virtualenv
 else
-    # temporary pyenv installation to get pypy-2.6 before container infra upgrade
+    # pyenv installation to get specified pypy version (Travis may only have older version(s))
     if [[ "${TOX_ENV}" == "pypy" ]]; then
         git clone https://github.com/yyuu/pyenv.git $PWD/.pyenv
         export PYENV_ROOT="$PWD/.pyenv"
         export PATH="$PYENV_ROOT/bin:$PATH"
         eval "$(pyenv init -)"
-        pyenv install pypy-2.6.0
-        pyenv global pypy-2.6.0
+        pyenv install "pypy-${PYPY_VERSION}"
+        pyenv global "pypy-${PYPY_VERSION}"
     fi
     pip install -U virtualenv
 fi