+++ HISTORY.rst
@@ -7,8 +7,8 @@ Upcoming
 ++++++++
 
 - CPython 3.5 support.
-- Support for cryptography>=1.0 on PyPy 2.6.
-- Travis CI testing for CPython 3.5 and PyPy 2.6.0.
+- Support for cryptography>=1.0 on PyPy >=2.6.
+- Travis CI testing for CPython 3.5 and PyPy 4.0.
 
 1.2.1 (2015-07-22)
 ++++++++++++++++++